backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.1
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
email-validator==2.1.0